cdk-agentcore/requirements.txt
aws-cdk-lib>=2.235.1,<3.0.0
aws-cdk.aws-bedrock-agentcore-alpha>=2.235.1a0,<3.0.0a0
constructs>=10.0.0,<11.0.0
